Tests now use the httptest package.
Tests were calling out to the live API, this generally isn't a good
idea. They were failing because the live endpoint had updated the
response from 8 items to 9.

I've modified the tests to now use the httptest package, the
checkidayURL is now just an unexported var instead of a const to allow us to
modify it within the same package.